requirements.txt
# requirements.txt for news scraping app
fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
python-dotenv
pytest
